# SubastasWeb/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Subastas",
        "description": "Documentación de la API de Subastas",
        "contact": {
            "email": "contacto@subastas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Obtener todos los clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "email",
                                    "contrasenia"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string"
                                    },
                                    "calificacion": {
                                        "type": "number"
                                    },
                                    "contrasenia": {
                                        "type": "string"
                                    },
                                    "latitud": {
                                        "type": "number"
                                    },
                                    "longitud": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "usuario_id": {
                                            "type": "integer"
                                        },
                                        "calificacion": {
                                            "type": "number"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "cedula": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "telefono": {
                                                    "type": "string"
                                                },
                                                "imagen": {
                                                    "type": "string"
                                                },
                                                "latitud": {
                                                    "type": "number"
                                                },
                                                "longitud": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/clientes/{usuario_id}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Obtener un cliente por su usuario_id",
                "operationId": "e291267aa80ff4384cc6590297522af8",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario asociado al cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Actualizar un cliente",
                "operationId": "c9f27624ed5819a0ce609517bdc49cef",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario asociado al cliente a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "email",
                                    "contrasenia"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string"
                                    },
                                    "calificacion": {
                                        "type": "number"
                                    },
                                    "latitud": {
                                        "type": "number"
                                    },
                                    "longitud": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Eliminar un cliente por usuario_id",
                "operationId": "71b1a2547ed0e7746edb731b8b78fcb5",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario asociado al cliente a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}